+++ src/matrix_rotation.c
@@ -6,7 +6,7 @@
 /*   By: wkorande <wkorande@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/11 17:47:57 by wkorande          #+#    #+#             */
-/*   Updated: 2019/12/11 17:50:04 by wkorande         ###   ########.fr       */
+/*   Updated: 2019/12/12 15:32:13 by wkorande         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@ t_mat4x4	create_rotation_matrix_x(float angle)
 
 	mat = create_identity_matrix();
 	mat.m[0][0] = 1.0f;
-	mat.m[1][1] = cosf(angle);
-	mat.m[1][2] = -sinf(angle);
-	mat.m[2][1] = sinf(angle);
-	mat.m[2][2] = cosf(angle);
+	mat.m[1][1] = cos(angle);
+	mat.m[1][2] = -sin(angle);
+	mat.m[2][1] = sin(angle);
+	mat.m[2][2] = cos(angle);
 	mat.m[3][3] = 1.0f;
 	return (mat);
 }
@@ -32,11 +32,11 @@ t_mat4x4	create_rotation_matrix_y(float angle)
 	t_mat4x4 mat;
 
 	mat = create_identity_matrix();
-	mat.m[0][0] = cosf(angle);
-	mat.m[0][2] = -sinf(angle);
+	mat.m[0][0] = cos(angle);
+	mat.m[0][2] = -sin(angle);
 	mat.m[1][1] = 1.0f;
-	mat.m[2][0] = sinf(angle);
-	mat.m[2][2] = cosf(angle);
+	mat.m[2][0] = sin(angle);
+	mat.m[2][2] = cos(angle);
 	return (mat);
 }
 
@@ -45,10 +45,10 @@ t_mat4x4	create_rotation_matrix_z(float angle)
 	t_mat4x4 mat;
 
 	mat = create_identity_matrix();
-	mat.m[0][0] = cosf(angle);
-	mat.m[0][1] = -sinf(angle);
-	mat.m[1][0] = sinf(angle);
-	mat.m[1][1] = cosf(angle);
+	mat.m[0][0] = cos(angle);
+	mat.m[0][1] = -sin(angle);
+	mat.m[1][0] = sin(angle);
+	mat.m[1][1] = cos(angle);
 	return (mat);
 }
